Move `generateDefaultNotebookName` to new file
